LayoutTests/fast/mutation/observe-exceptions-expected.txt

Test that WebKitMutationObserver.observe throws exceptions appropriately

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS observer.observe() threw exception TypeError: Not enough arguments.
PASS observer.observe(null) threw exception TypeError: Not enough arguments.
PASS observer.observe(undefined) threw exception TypeError: Not enough arguments.
PASS observer.observe(document.body) threw exception TypeError: Not enough arguments.
PASS observer.observe(document.body, null) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS observer.observe(document.body, undefined) threw exception Error: TYPE_MISMATCH_ERR: DOM Exception 17.
PASS successfullyParsed is true

TEST COMPLETE
